Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
psycopg[binary]
motor
pydantic[email]
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
httpx
bson